requirements.txt
# Core Animation Library
manim>=0.18.0

# GUI Framework (Includes QtMultimedia)
PySide6>=6.4.0

# Google Gemini AI SDK
google-genai>=0.3.0

# Audio Processing
# Note: You must also have FFmpeg installed on your system PATH
pydub>=0.25.0

# API Requests (For the LaTeX preview feature)
requests>=2.31.0

# Numerical Operations
numpy>=1.24.0
